Update package-lock.json and package.json
This commit updates the `package-lock.json` and `package.json` files. It changes the name of the project to "tecfancy-cli" and updates the devDependencies with new versions of `@types/node` and `typescript`. It also adds new scripts to the `package.json` file for building and watching the project. Additionally, a new `tsconfig.json` file is added with compiler options.